투사체 기능 작업

+++ Assets/2DPlatformerRND/01_Scripts/Managers/SkillResourceData.cs
@@ -13,22 +13,22 @@ namespace PahlBit
         public readonly string DisplayName;
         public readonly string Desc;
 
-        public readonly double Attack;
-        public readonly double AttackPerLv;
-        public readonly double ManaUse;
-        public readonly double ManaUsePerLv;
-        public readonly double Cooltime;
-        public readonly double CooltimeDownPerLv;
-        public readonly double ProjectileCount;
-        public readonly double ProjectileCountPerLv;
-        public readonly double AttackRange;
-        public readonly double AttackRangePerLv;
-        public readonly double SplashRange;
-        public readonly double SplashRangePerLv;
-        public readonly double Duration;
-        public readonly double DurationPerLv;
-        public readonly double Interval;
-        public readonly double IntervalPerLv;
+        public readonly float Attack;
+        public readonly float AttackPerLv;
+        public readonly float ManaUse;
+        public readonly float ManaUsePerLv;
+        public readonly float Cooltime;
+        public readonly float CooltimeDownPerLv;
+        public readonly float ProjectileCount;
+        public readonly float ProjectileCountPerLv;
+        public readonly float AttackRange;
+        public readonly float AttackRangePerLv;
+        public readonly float SplashRange;
+        public readonly float SplashRangePerLv;
+        public readonly float Duration;
+        public readonly float DurationPerLv;
+        public readonly float Interval;
+        public readonly float IntervalPerLv;
 
         public int RowIndex { get; set; } // 데이터데이블상에 존재하는 순서
         public long ID { get { return ToID(SkillID); } } // 데이터 접근을 위한 id값
